Refactor ETC decompression and move it to it's own package. The next PR will add the capability of compression with CGo that would slow the builds. Therefore similar to ASTC, ETC should have it's own package.
