+++ core/image/etc1.go
@@ -19,10 +19,6 @@ import (
 	"github.com/google/gapid/core/stream"
 )
 
-var (
-	ETC1_RGB_U8_NORM = NewETC1_RGB_U8_NORM("ETC1_RGB_U8_NORM")
-)
-
 // NewETC1_RGB_U8_NORM returns a format representing the ETC1_RGB8 block texture
 // compression format.
 func NewETC1_RGB_U8_NORM(name string) *Format {
@@ -40,13 +36,4 @@ func (f *FmtETC1_RGB_U8_NORM) check(data []byte, w, h, d int) error {
 }
 func (*FmtETC1_RGB_U8_NORM) channels() stream.Channels {
 	return stream.Channels{stream.Channel_Red, stream.Channel_Green, stream.Channel_Blue}
-}
-
-func init() {
-	RegisterConverter(ETC1_RGB_U8_NORM, RGB_U8_NORM, func(src []byte, w, h, d int) ([]byte, error) {
-		return Convert(src, w, h, d, ETC2_RGB_U8_NORM, RGB_U8_NORM)
-	})
-	RegisterConverter(ETC1_RGB_U8_NORM, RGBA_U8_NORM, func(src []byte, w, h, d int) ([]byte, error) {
-		return Convert(src, w, h, d, ETC2_RGB_U8_NORM, RGBA_U8_NORM)
-	})
 }
